corebootPkg/Sec/X64: Align SecEntry to page size
We need to setup page tables on X64 which must be page aligned.
This never quite worked since the final step (adding SecMain.efi to
the firmware volume) linked the file to some non-4K alignment, so
enforce it.

Requires latest cbfstool with support for alignment related padding
data in FV.

Signed-off-by: Patrick Georgi <patrick@georgi-clan.de>